(PC-13216)[PRO] chore: renamve test files
